Move highlight logic out of EditorProvider
